requirements.txt
numpy
pandas
scikit-learn
matplotlib
xai-sdk
tweepy
python-dotenv
Flask